Small fixes according to @nrc's comment
