+++ src/storage/txn/commands/mod.rs
@@ -1,4 +1,7 @@
 // Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+//! Commands used in the transaction system
+#[macro_use]
+mod macros;
 mod acquire_pessimistic_lock;
 mod check_txn_status;
 mod cleanup;
@@ -316,138 +319,6 @@ pub trait CommandExt: Display {
     fn write_bytes(&self) -> usize;
 
     fn gen_lock(&self, _latches: &Latches) -> latch::Lock;
-}
-
-#[macro_export]
-macro_rules! ctx {
-    () => {
-        fn get_ctx(&self) -> &Context {
-            &self.ctx
-        }
-        fn get_ctx_mut(&mut self) -> &mut Context {
-            &mut self.ctx
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! command {
-    (
-        $(#[$outer_doc: meta])*
-        $cmd: ident:
-            cmd_ty => $cmd_ty: ty,
-            display => $format_str: expr, ($($fields: ident$(.$sub_field:ident)?),*),
-            content => {
-                $($(#[$inner_doc:meta])* $arg: ident : $arg_ty: ty,)*
-            }
-    ) => {
-        $(#[$outer_doc])*
-        pub struct $cmd {
-            pub ctx: Context,
-            $($(#[$inner_doc])* pub $arg: $arg_ty,)*
-        }
-
-        impl $cmd {
-            pub fn new(
-                $($arg: $arg_ty,)*
-                ctx: Context,
-            ) -> TypedCommand<$cmd_ty> {
-                Command::$cmd($cmd {
-                        ctx,
-                        $($arg,)*
-                }).into()
-            }
-        }
-
-        impl Display for $cmd {
-            fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-                write!(
-                    f,
-                    $format_str,
-                    $(
-                        self.$fields$(.$sub_field())?,
-                    )*
-                )
-            }
-        }
-
-        impl Debug for $cmd {
-            fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-                write!(f, "{}", self)
-            }
-        }
-    }
-}
-
-#[macro_export]
-macro_rules! ts {
-    ($ts:ident) => {
-        fn ts(&self) -> TimeStamp {
-            self.$ts
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! tag {
-    ($tag:ident) => {
-        fn tag(&self) -> metrics::CommandKind {
-            metrics::CommandKind::$tag
-        }
-
-        fn incr_cmd_metric(&self) {
-            KV_COMMAND_COUNTER_VEC_STATIC.$tag.inc();
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! write_bytes {
-    ($field: ident) => {
-        fn write_bytes(&self) -> usize {
-            self.$field.as_encoded().len()
-        }
-    };
-    ($field: ident: multiple) => {
-        fn write_bytes(&self) -> usize {
-            self.$field.iter().map(|x| x.as_encoded().len()).sum()
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! gen_lock {
-    (empty) => {
-        fn gen_lock(&self, _latches: &Latches) -> latch::Lock {
-            latch::Lock::new(vec![])
-        }
-    };
-    ($field: ident) => {
-        fn gen_lock(&self, latches: &Latches) -> latch::Lock {
-            latches.gen_lock(iter::once(&self.$field))
-        }
-    };
-    ($field: ident: multiple) => {
-        fn gen_lock(&self, latches: &Latches) -> latch::Lock {
-            latches.gen_lock(&self.$field)
-        }
-    };
-    ($field: ident: multiple$transform: tt) => {
-        fn gen_lock(&self, latches: &Latches) -> latch::Lock {
-            #![allow(unused_parens)]
-            let keys = self.$field.iter().map($transform);
-            latches.gen_lock(keys)
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! command_method {
-    ($name:ident, $return_ty: ty, $value: expr) => {
-        fn $name(&self) -> $return_ty {
-            $value
-        }
-    };
 }
 
 impl Command {
