+++ src/storage/txn/commands/pause.rs
@@ -1,9 +1,6 @@
-use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
 use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
-use crate::storage::txn::latch::{self, Latches};
-use crate::storage::Context;
-use crate::{command, ctx, gen_lock, tag, write_bytes};
-use std::fmt::{self, Debug, Display, Formatter};
 use txn_types::Key;
 
 command! {
