+++ src/storage/txn/commands/prewrite.rs
@@ -1,10 +1,8 @@
-use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
 use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
-use crate::storage::txn::latch::{self, Latches};
-use crate::storage::{Context, Result, TimeStamp};
-use crate::{command, ctx, gen_lock, tag, ts};
-use std::fmt::{self, Debug, Display, Formatter};
-use txn_types::Mutation;
+use crate::storage::{Context, Result};
+use txn_types::{Mutation, TimeStamp};
 
 command! {
     /// The prewrite phase of a transaction. The first phase of 2PC.
