+++ src/storage/txn/commands/resolve_lock.rs
@@ -1,11 +1,8 @@
-use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
 use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
-use crate::storage::txn::latch::{self, Latches};
-use crate::storage::{Context, TimeStamp};
-use crate::{command, command_method, ctx, gen_lock, tag};
-use std::fmt::{self, Debug, Display, Formatter};
 use tikv_util::collections::HashMap;
-use txn_types::{Key, Lock};
+use txn_types::{Key, Lock, TimeStamp};
 
 command! {
     /// Resolve locks according to `txn_status`.
