+++ src/storage/txn/commands/scan_lock.rs
@@ -1,11 +1,8 @@
-use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
 use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
-use crate::storage::txn::latch::{self, Latches};
 use crate::storage::txn::LockInfo;
-use crate::storage::{Context, TimeStamp};
-use crate::{command, command_method, ctx, gen_lock, tag, ts};
-use std::fmt::{self, Debug, Display, Formatter};
-use txn_types::Key;
+use txn_types::{Key, TimeStamp};
 
 command! {
     /// Scan locks from `start_key`, and find all locks whose timestamp is before `max_ts`.
