+++ src/storage/txn/commands/txn_heart_beat.rs
@@ -1,11 +1,8 @@
-use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
 use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
-use crate::storage::txn::latch::{self, Latches};
-use crate::storage::{Context, TimeStamp, TxnStatus};
-use crate::{command, ctx, gen_lock, tag, ts, write_bytes};
-use std::fmt::{self, Debug, Display, Formatter};
-use std::iter;
-use txn_types::Key;
+use crate::storage::TxnStatus;
+use txn_types::{Key, TimeStamp};
 
 command! {
     /// Heart beat of a transaction. It enlarges the primary lock's TTL.
